# Represent Expr c as Seq (Term c) instead of [Term c]