`sum`  adds items in the left-associative way, which causes performance degradation.

In a case of `toyconvert +RTS -p -RTS normalized-Generated_notes_20140709_1000_0.01-000.opb -o normalized-Generated_notes_20140709_1000_0.01-000.lp` (`normalized-Generated_notes_20140709_1000_0.01-000.opb` is from https://www.cril.univ-artois.fr/PB24/benchs/normalized-PB24.tar), the result is disastrous.

This PR fixes the problem by replacing the internal representation of `Expr` from lists to `Seq`. Compatibility is kept using pattern synonyms.

Before:

```
	total time  =      200.54 secs   (200542 ticks @ 1000 us, 1 processor)
	total alloc = 595,219,815,488 bytes  (excludes profiling overheads)

COST CENTRE MODULE                        SRC                                            %time %alloc

+           Numeric.Optimization.MIP.Base src/Numeric/Optimization/MIP/Base.hs:213:3-37   96.7   97.2
```

After:

```
	total time  =        3.81 secs   (3805 ticks @ 1000 us, 1 processor)
	total alloc = 16,827,186,784 bytes  (excludes profiling overheads)
```